+++ internal/app/storage/storage.go
@@ -61,7 +61,7 @@ type Storage interface {
 		ctx context.Context,
 		tx pgx.Tx,
 		userID int,
-		sum int,
+		sum float32,
 		orderID int,
 	) error
